moving to homebridge typed system
