+++ src/interface/config.ts
@@ -1,9 +1,11 @@
+import { PlatformConfig } from 'homebridge';
+
 export interface BondConfig {
   ip_address: string;
   token: string;
 }
 
-export interface BondPlatformConfig {
+export interface BondPlatformConfig extends PlatformConfig {
   bonds: BondConfig[];
   debug: boolean;
   include_dimmer: boolean;
